crud color + loai os

+++ src/main/java/com/example/world_phone/dto/respone/attribute/os/OsRespone.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class OsRespone {
     private Long id;
     private String name;
-    private int loai;
+    private Long loai;
 }
